+++ pkg/skaffold/runner/dev.go
@@ -19,18 +19,8 @@ package runner
 import (
 	"context"
 	"fmt"
-	kubernetesclient "github.com/GoogleContainerTools/skaffold/v2/pkg/skaffold/kubernetes/client"
-	"github.com/GoogleContainerTools/skaffold/v2/proto/filedownload"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 	"io"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"net"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/GoogleContainerTools/skaffold/v2/pkg/skaffold/constants"
@@ -118,7 +108,6 @@ func (r *SkaffoldRunner) doDev(ctx context.Context, out io.Writer) error {
 	}
 
 	var bRes []graph.Artifact
-	artifactsToRebuild := r.changeSet.NeedsRebuild()
 	if needsBuild {
 		childCtx, endTrace := instrumentation.StartTrace(ctx, "doDev_needsBuild")
 		event.ResetStateOnBuild()
@@ -214,102 +203,8 @@ func (r *SkaffoldRunner) doDev(ctx context.Context, out io.Writer) error {
 		if err := r.deployer.GetDebugger().Start(childCtx); err != nil {
 			log.Entry(ctx).Warnf("failed to start debugger: %v", err)
 		}
-		builds := r.Builds
-		go func() {
 
-			kClient, err2 := kubernetesclient.Client("minikube")
-			if err2 != nil {
-				fmt.Println(err2)
-			}
-			list, _ := kClient.CoreV1().Pods("default").List(ctx, metav1.ListOptions{})
-			for _, p := range list.Items {
-				for _, c := range p.Spec.Containers {
-					if ds := getDownstreamSync(ctx, artifactsToRebuild, builds, c.Image); ds != nil {
-						pr, pw := io.Pipe()
-						gr, gw := io.Pipe()
-						command := exec.CommandContext(ctx, "kubectl", "exec", "-it", "pods/"+p.Name, "-c", c.Name, "--", "/abccc/app-connect")
-						command.Stdout = pw
-						command.Stdin = gr
-						err = command.Start()
-						fmt.Println(command.Args)
-						if err != nil {
-							fmt.Println("failed to connect the remote ")
-							fmt.Println(err)
-						}
-						conn, err := grpc.DialContext(ctx, "", grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithContextDialer(func(ctx context.Context, s string) (net.Conn, error) {
-							return &Conn{pr, gw}, nil
-						}))
-						if err != nil {
-							fmt.Println(err)
-						}
-
-						client := filedownload.NewFileServiceClient(conn)
-
-						watch, err := client.Watch(ctx, &filedownload.FileWatchRequest{})
-						fmt.Println("watch called")
-						if err != nil {
-							fmt.Println(err)
-						}
-
-						go func() {
-							for {
-								recv, err2 := watch.Recv()
-								if err2 != nil {
-									fmt.Println(err2)
-									return
-								}
-
-								fmt.Println(recv)
-								for _, entry := range ds.Entry {
-
-									if !MatchDir(entry.Src, recv.Path) {
-										continue
-									}
-									rel, err2 := filepath.Rel(entry.Src, recv.Path)
-									if err2 != nil {
-										fmt.Println(err2)
-										continue
-									}
-
-									t := filepath.Join(entry.Dst, rel)
-									if v, ok := filemon.SyncedHash[t]; ok {
-										if v == recv.MD5Hash {
-											fmt.Println("File already synced.")
-											continue
-										}
-									}
-									fmt.Println("Syncing remote file to local")
-									filemon.SyncedHash[t] = recv.MD5Hash
-									fmt.Println(filemon.SyncedHash)
-
-									file, err2 := client.DownloadFile(context.Background(), &filedownload.DownloadRequest{Path: recv.Path})
-									if err2 != nil {
-										fmt.Println(err2)
-									}
-
-									os.MkdirAll(filepath.Dir(t), 0755)
-									create, err2 := os.Create(t)
-									if err2 != nil {
-										fmt.Println("failed to create")
-										fmt.Println(err2)
-									}
-									for {
-										response, err2 := file.Recv()
-										if err2 == io.EOF {
-											break
-										}
-										create.Write(response.Chunk)
-									}
-									create.Close()
-
-								}
-
-							}
-						}()
-					}
-				}
-			}
-		}()
+		r.deployer.GetDownloader().Start(ctx, bRes)
 
 		endTrace()
 	}
@@ -429,101 +324,7 @@ func (r *SkaffoldRunner) Dev(ctx context.Context, out io.Writer, artifacts []*la
 	}
 
 	builds := r.Builds
-	go func() {
-
-		kClient, err2 := kubernetesclient.Client("minikube")
-		if err2 != nil {
-			fmt.Println(err2)
-		}
-		list, _ := kClient.CoreV1().Pods("default").List(ctx, metav1.ListOptions{})
-		for _, p := range list.Items {
-			for _, c := range p.Spec.Containers {
-				if ds := getDownstreamSync(ctx, artifacts, builds, c.Image); ds != nil {
-					pr, pw := io.Pipe()
-					gr, gw := io.Pipe()
-					command := exec.CommandContext(ctx, "kubectl", "exec", "-it", "pods/"+p.Name, "-c", c.Name, "--", "/abccc/app-connect")
-					command.Stdout = pw
-					command.Stdin = gr
-					err = command.Start()
-					fmt.Println(command.Args)
-					if err != nil {
-						fmt.Println("failed to connect the remote ")
-						fmt.Println(err)
-					}
-					conn, err := grpc.DialContext(ctx, "", grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithContextDialer(func(ctx context.Context, s string) (net.Conn, error) {
-						return &Conn{pr, gw}, nil
-					}))
-					if err != nil {
-						fmt.Println(err)
-					}
-
-					client := filedownload.NewFileServiceClient(conn)
-
-					watch, err := client.Watch(ctx, &filedownload.FileWatchRequest{})
-					fmt.Println("watch called")
-					if err != nil {
-						fmt.Println(err)
-					}
-
-					go func() {
-						for {
-							recv, err2 := watch.Recv()
-							if err2 != nil {
-								fmt.Println(err2)
-								return
-							}
-
-							fmt.Println(recv)
-							for _, entry := range ds.Entry {
-
-								if !MatchDir(entry.Src, recv.Path) {
-									continue
-								}
-								rel, err2 := filepath.Rel(entry.Src, recv.Path)
-								if err2 != nil {
-									fmt.Println(err2)
-									continue
-								}
-
-								t := filepath.Join(entry.Dst, rel)
-								if v, ok := filemon.SyncedHash[t]; ok {
-									if v == recv.MD5Hash {
-										fmt.Println("File already synced.")
-										continue
-									}
-								}
-								fmt.Println("Syncing remote file to local")
-								filemon.SyncedHash[t] = recv.MD5Hash
-								fmt.Println(filemon.SyncedHash)
-
-								file, err2 := client.DownloadFile(context.Background(), &filedownload.DownloadRequest{Path: recv.Path})
-								if err2 != nil {
-									fmt.Println(err2)
-								}
-
-								os.MkdirAll(filepath.Dir(t), 0755)
-								create, err2 := os.Create(t)
-								if err2 != nil {
-									fmt.Println("failed to create")
-									fmt.Println(err2)
-								}
-								for {
-									response, err2 := file.Recv()
-									if err2 == io.EOF {
-										break
-									}
-									create.Write(response.Chunk)
-								}
-								create.Close()
-
-							}
-
-						}
-					}()
-				}
-			}
-		}
-	}()
+	r.deployer.GetDownloader().Start(ctx, builds)
 
 	g := getTransposeGraph(artifacts)
 	// Watch artifacts
@@ -651,39 +452,4 @@ func getTransposeGraph(artifacts []*latest.Artifact) devGraph {
 		}
 	}
 	return g
-}
-
-func getDownstreamSync(ctx context.Context, artifacts []*latest.Artifact, builds []graph.Artifact, containerImage string) *latest.DownstreamSync {
-
-	g := graph.ToArtifactGraph(artifacts)
-
-	for _, b := range builds {
-		if b.Tag != containerImage {
-			continue
-		}
-		if v, ok := g[b.ImageName]; ok {
-			return v.DownstreamSync
-		}
-	}
-	return nil
-}
-
-func MatchDir(targetDir string, changedDir string) bool {
-	if targetDir == "." {
-		return true
-	}
-	list1 := strings.Split(targetDir, string(os.PathSeparator))
-	list2 := strings.Split(changedDir, string(os.PathSeparator))
-	fmt.Println(list1)
-	fmt.Println(list2)
-
-	if len(list1) > len(list2) {
-		return false
-	}
-	for i, ele := range list1 {
-		if list2[i] != ele {
-			return false
-		}
-	}
-	return true
 }
